refactor: address command localization review
