+++ src/commands/index.ts
@@ -44,120 +44,6 @@ import {
   type SystemCommands,
 } from './system/runtime/index.ts';
 
-export type { ScreenshotCommandResult } from './capture/runtime/screenshot.ts';
-export type {
-  DiffScreenshotCommandOptions,
-  DiffScreenshotCommandResult,
-  LiveScreenshotInputRef,
-} from './capture/runtime/diff-screenshot.ts';
-export type {
-  DiffSnapshotCommandResult,
-  SnapshotCommandResult,
-  SnapshotDiffLine,
-  SnapshotDiffSummary,
-} from './capture/runtime/snapshot.ts';
-export type {
-  ElementTarget,
-  FindReadCommandOptions,
-  FindReadCommandResult,
-  GetAttrsCommandOptions,
-  GetCommandOptions,
-  GetCommandResult,
-  GetTextCommandOptions,
-  IsCommandOptions,
-  IsCommandResult,
-  IsSelectorCommandOptions,
-  RefTarget,
-  ResolvedTarget,
-  SelectorSnapshotOptions,
-  SelectorTarget,
-  WaitCommandOptions,
-  WaitCommandResult,
-  WaitForTextCommandOptions,
-} from './interaction/runtime/selector-read.ts';
-export type {
-  ClickCommandOptions,
-  FillCommandOptions,
-  FillCommandResult,
-  FocusCommandOptions,
-  FocusCommandResult,
-  InteractionTarget,
-  LongPressCommandOptions,
-  LongPressCommandResult,
-  PinchCommandOptions,
-  PinchCommandResult,
-  PointTarget,
-  PressCommandOptions,
-  PressCommandResult,
-  ResolvedInteractionTarget,
-  ScrollCommandOptions,
-  ScrollCommandResult,
-  ScrollTarget,
-  SwipeCommandOptions,
-  SwipeCommandResult,
-  SwipeOptions,
-  TypeTextCommandOptions,
-  TypeTextCommandResult,
-} from './interaction/runtime/interactions.ts';
-export type {
-  AppPushInput,
-  CloseAppCommandOptions,
-  CloseAppCommandResult,
-  GetAppStateCommandOptions,
-  GetAppStateCommandResult,
-  ListAppsCommandOptions,
-  ListAppsCommandResult,
-  OpenAppCommandOptions,
-  OpenAppCommandResult,
-  PushAppCommandOptions,
-  PushAppCommandResult,
-  TriggerAppEventCommandOptions,
-  TriggerAppEventCommandResult,
-} from './management/runtime/apps.ts';
-export type {
-  AdminBootCommandOptions,
-  AdminBootCommandResult,
-  AdminDevicesCommandOptions,
-  AdminDevicesCommandResult,
-  AdminInstallCommandOptions,
-  AdminInstallCommandResult,
-  AdminInstallFromSourceCommandOptions,
-  AdminReinstallCommandOptions,
-  AdminShutdownCommandOptions,
-  AdminShutdownCommandResult,
-} from './management/runtime/admin.ts';
-export type {
-  DiagnosticsLogsCommandOptions,
-  DiagnosticsLogsCommandResult,
-  DiagnosticsNetworkCommandOptions,
-  DiagnosticsNetworkCommandResult,
-  DiagnosticsPerfCommandOptions,
-  DiagnosticsPerfCommandResult,
-} from './observability/runtime/diagnostics.ts';
-export type {
-  RecordingRecordCommandOptions,
-  RecordingRecordCommandResult,
-  RecordingTraceCommandOptions,
-  RecordingTraceCommandResult,
-} from './recording/runtime/recording.ts';
-export type {
-  SystemAlertCommandOptions,
-  SystemAlertCommandResult,
-  SystemAppSwitcherCommandOptions,
-  SystemAppSwitcherCommandResult,
-  SystemBackCommandOptions,
-  SystemBackCommandResult,
-  SystemClipboardCommandOptions,
-  SystemClipboardCommandResult,
-  SystemHomeCommandOptions,
-  SystemHomeCommandResult,
-  SystemKeyboardCommandOptions,
-  SystemKeyboardCommandResult,
-  SystemRotateCommandOptions,
-  SystemRotateCommandResult,
-  SystemSettingsCommandOptions,
-  SystemSettingsCommandResult,
-} from './system/runtime/system.ts';
 export { ref, selector } from './interaction/runtime/selector-read.ts';
 
 export type {
